backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-jose[cryptography]
python-multipart
pydantic
python-dotenv
passlib[bcrypt]
